Install and configure @websanova/vue-auth library
This is the library responsible for handling authentication on the client side. It injects an $auth object which provides a handful of helper functions such as register() which handles user registration, login() which handles user login, user() which provides access to the current user data, logout() with handles logout, and a couple of other functions.